refactor(ai-packages): normalize new package filenames
