+++ packages/ai-core/src/core/providers/__tests__/openai-image-model.test.ts
