+++ packages/ai-sdk-provider/src/index.ts
@@ -1,3 +1,3 @@
 export * from './cherryin-provider'
 export * from './openai-compatible-reranking-model'
-export * from './reasoningModelTransform'
+export * from './reasoning-model-transform'
