+++ packages/ai-sdk-provider/src/reasoning-model-transform.ts
